feat: move sendDrip function into utilities
