+++ .env.example
@@ -7,7 +7,7 @@ DEFAULT_RPC=https://rpc-df.xdaichain.com/
 # DF_WEBSERVER_URL=http://localhost:3000
 
 # If you want to use a faucet, put url here
-FAUCET_URL=http://localhost:3000
+DFDAO_WEBSERVER_URL=http://localhost:3003
 
 # If you want to use just the Twitter integration, add this:
 DF_TWITTER_URL=https://api.zkga.me
