+++ src/Backend/Network/Blockchain.ts
@@ -62,7 +62,7 @@ export function getEthConnection(): Promise<EthConnection> {
   console.log(`is production network: ${isProdNetwork}`);
   console.log(`client build: ${process.env.NODE_ENV}`);
   console.log(`webserver url: ${process.env.DF_WEBSERVER_URL}`);
-  console.log(`faucet url: ${process.env.FAUCET_URL}`);
+  console.log(`faucet url: ${process.env.DFDAO_WEBSERVER_URL}`);
 
   return createEthConnection(url);
 }
